feat(signing): Add support for signing each change.
In p2p and offline systems it's often necessary to sign each change so that the "actor" noted in the
change can be validated at any time by any participant. Rather than add a bunch of crypto and strong
opinions about *how* to sign I added support for an additional function in the change method that
can be used to return a signature for any change object. This will allow people to implement their
own signing and validation methods without any additional features in automerge.