Replaced generators.
Refactored generators to regular functions to eliminate the necessity of
a generator polyfill, which causes a lot of unfortunate distribution
overhead (i.e. the use of Babel's transform-runtime, which I have been
unable to integrate successfuly).  It also reduces the size of the
generated bundle by ~260k (over the previous use of babel-polyfill,
which I have learned is a poor choice to use in libraries).