Downgrade uuid dependency to solve loading problem
Fixes #104.

In the webpack-packaged output, the file lib/rng.js is replaced with
lib/rng-browser.js (due to the 'browser' section declaration in
package.json). uuid version 3.2.1 checks for the existence of a global
`crypto` object, and then does `crypto.getRandomValues.bind(crypto)`
without checking that `crypto.getRandomValues` exists. Turns out that in
Node, `crypto` exists but `crypto.getRandomValues` doesn't.

Looks like uuid version 3.3.0 fixes this issue, but it hasn't yet been
released on npm.

More fundamentally, the issue here is probably that we're using a
browser-targeted packaged version in Node. It would be better to use the
uuid package's lib/rng.js in Node, since it uses a better random number
generator. But I'm not sure how to correctly set up the packaging.