Require that the change message must be a string
At the moment, the second argument to `Automerge.change()` can be not
just a message describing the change, but it can in fact be any
JavaScript object, and it is encoded as JSON in the serialised change
data.

I am proposing changing the behaviour such that `Automerge.change()`
only accepts a string as change message, and throws an error otherwise.
This will pave the way towards supporting a richer API in the future,
e.g. using an options object. This change also helps with the binary
encoding of changes.

If you want to continue using the change message for arbitrary
structured metadata can do so by doing `JSON.stringify` and passing in
the result as change message.