Move sequence number assignment for undo/redo to frontend
Added Frontend.undo() and Frontend.redo(), which increment the sequence
number and enqueue a request to be sent to the backend. The backend uses
the actor and seq from the request, and adds in the ops. This ensures
that only the frontend ever increment seq. Moreover, it removes the main
reason why the backend needs to know the actorId. Going to remove the
backend's dependency on actorId in the next commit.