Separate clock and deps in patch metadata
Previously, the conflation of deps and clock was bit of a mess: deps
contained the minimal dependencies (not those actors that are
transitively included in one of the other dependencies), except for the
local actor, which was always included. This change sends the plain deps
from the backend to the frontend (without special-casing the local
actor). In order to tell the frontend what its sequence number should be
(after loading from disk, for example), in principle it would suffice to
include just that sequence in the patch; however, that requires the
backend to know what the frontend's actorId is. Since there is now
nothing else in the backend that requires knowledge of the local
actorId, it would be nice to remove the actorId requirement. To that
end, we send the entire vector clock in the patch (in addition to deps),
and the frontend can pick out the component corresponding to its
actorId.