Backend no longer needs to know actorId
Thanks to the last few commits, the backend no longer needs to know what
actorId is the local one. This is nice for a few reasons: it makes the
backend more generic, it removes the danger of accidentally
misconfiguring (e.g. setting frontend and backend to different actors),
and it allows read-only backends to exist without requiring any actorId
at all. This is useful in a Hypermerge context, where actorIds are
public keys, and so actorId creation is fairly heavyweight; in this
context it is desirable to defer actorId creation until the first change
is made.