use strict mode everywhere
The original code here suggests that in a browser,
modifying a frozen object will fail silently, whereas
in node it will throw an error. This test was failing
for me in browser testing, so I dug into it and it
turns out the difference isn't node vs browser
but strict mode vs not. (In mocha.opts we had
"use strict" so node tests were always running
in strict mode. TypeScript was also compiling in
strict mode.)

Since we have control over whether we're in strict 
mode or not, it makes more sense to just turn it on
everywhere and only test for that scenario.

screw it, just use strict mode everywhere