Exception on modifying frozen object is optional
In Node v12.3.1 and v10.16.0, assigning or deleting a property of a
frozen object no longer throws an exception like it did in previous
versions. Rather than trying to split the test by version number, I
figured it was easier to just remove the assertion that an exception is
thrown (and just check that the object remains unmodified).