Detect the current platform's behaviour for throwing on frozen object modification for the tests
Different platforms sadly treat modifying a frozen object differently, so this test has to test different behaviour depending on the platform. Before this change, it hardcoded which behaviour to expect when based on the presence of the window object, but new node versions changed their behaviour to match that of the browsers, breaking this check. Instead of a hardcoded check, lets just detect the current platform's behaviour outside the test, and then make sure that automerge matches that behaviour.

This fixes the tests to run cleanly on node v10.x and v11.x