Fix stack overflow error on large changes
Fixes #202. The stack overflow is not due to unbounded recursion, but
rather because of a method's varargs argument list being too long
(presumably each argument takes up some space on the stack). In
particular, there are several places where we are calling Array.push
with an argument list generated using the spread operator in order to
concatenate arrays. This commit replaces them with for loops, which also
happen to be faster: https://jsperf.com/pushing-large-arrays