Fixed undo/redo
Did this by reintroducing a 'link' operation, which assigns an existing
object as a child of another object. The alternative for undoing the
deletion of a child object would have been to generate operations to
materialise a new object (with new objectId), but that seemed ugly too.
When we introduce a move operation, we will have something very much
like 'link' anyway, so it doesn't seem too bad to re-introduce it now.