Use sequential version numbers instead of deps
On change requests (from frontend to backend) and patches (from backend
to frontend), we now use a scalar version number instead of the `deps`
vector clock. The version is incremented on each incremental change
processed by the backend (both local and remote changes). The frontend
does not increment the version (instead, the frontend increments the
sequence number `seq` on local changes).

This lays the groundwork for the next commit...