Make index-to-elemId translation work with frontend/backend concurrency
The problem with translating list indexes to list elemIds in the backend
is that when the backend receives a change request from the frontend,
the backend may have already applied a remote change that had not yet
been seen by the frontend at the time it made the change request, and
that remote change may have affected the indexes by inserting or
removing earlier list elements.

This patch fixes the issue by having the backend explicitly keep track
of the state that the frontend is in. As long as there are no remote
changes, this is easy: the latest backend state equals the frontend
state. When a remote change comes in, we keep separate copies of the
OpSet for the version of the document that the frontend is using; these
OpSets incorporate local changes but omit remote changes, which means
they should be in the same state as the frontend. These OpSets are then
used for translating indexes, and discarded once the frontend has
received the patches for the remote changes.