Avoid speculative application of local patches
This is a surprisingly simple resolution to a long-standing issue; I
only just realised that this simple solution existed. The problem is:
when using separate frontend and backend, the frontend can make a change
request, and concurrently the backend can receive a remote change. The
frontend will then receive the patch corresponding to the remote change
before it receives the confirmation patch for its own local change. If
the frontend applies the patch for the remote change to the base
document, its own local change would temporarily disappear. But if the
frontend re-applies the patch for its local change on top of the remote
patch, the local patch may no longer apply correctly, especially if the
remote patch changed list indexes.

In the past we had a transformRequest() function that performed a kind
of basic Operational Transformation to the local patch in order to
account for changes to list indexes. But there were edge cases that this
function did not handle correctly, so I removed it since it wasn't
really making things better.

This commit makes a much simpler solution: if the frontend receives a
patch for a remote change while a local change is still in flight, that
remote patch does not take effect immediately, but rather it is queued
up. All patches take effect once the queue of in-flight change requests
is drained. This ensures all patches apply cleanly (since all patches
are applied in the order determined by the backend). Much simpler!