Remove aggregated patch from Context
Previously, the Context needed to generate two patches: one to apply
immediately (so that subsequent commands within the same change callback
see the effect of the changes), and one that aggregates all changes that
were made within the callback. As of the previous commit, the latter is
no longer needed. Nice simplification.