Remove the list of columns from Automerge.Table
Since #236 the Automerge.Table type no longer allows rows to be added as
an array of values, which are then mapped to column names. Now that this
feature is removed, nothing in Automerge is actually using the list of
columns that is currently required by the Automerge.Table constructor.
The columns are still saved, but since they do not enforce any schema,
they are purely advisory.

Hopefully one day we can have proper schema support in Automerge, but
the current half-hearted implementation is not really helping us get
there. Therefore I think it is best to just remove this list of columns
feature.