Remove set() method from Automerge.Table API
It is not needed, since add() and remove() handle changes to the set of
rows, and properties of row objects can be updated directly without
requiring any set() operation at the table level.

Looking at the code, I now realise that set() was not intended to be
part of the public API at all: it is called while applying a patch, and
if a user calls it, it does not generate any operations in the change
context. Hence I renamed it to start with an underscore (like `_clone()`
and `_freeze()`), and added a warning to the comment.