Allow the backend to be mutable
Detect if users are referencing a backend object whose state has been
updated. While the current immutable backend allows multiple references
to old and current versions of the backend state, this is not the case
when the backend becomes mutable. So it's best to complain loudly if the
user does try to access an old backend state.

Fixed the tests that were in fact referring to outdated backend states.