Fill in pred fields based on state seen by frontend
If a remote change updated the value of a property, and concurrently
with the backend processing that change the frontend requested a change
that updated the same property, then the pred field of the local
operation would get filled in based on the state after applying the
remote change. This commit fixes the behaviour: now the pred field is
filled in based on the state seen by the frontend when it requested the
change (in this case, that state would not include the remote change, so
the desired outcome is a conflict rather than overwriting the remote
value).

Thanks to @orionz for the bug report.