Use hashes instead of actorId+seqNo for dependencies
As discussed in #200. This makes Automerge more like Git, in that the
entire editing history is captured in a single hash, and all prior
changes are reachable through the hash graph. By using a
cryptographically strong hash function, it becomes much easier to
authenticate history and check whether two replicas really are in sync
(guaranteeing that any inconsistencies will be detected).

The current implementation throws an exception if a sequence number is
reused by the same actor with a different change hash. In the future it
would be nice to gracefully handle that situation and allow clean
merging.

Also, so far this implementation ignores the problem of how network
protocols will figure out which changes replicas need to send to each
other in order to converge. Doing this based on hash chains is annoying,
as it may require several round-trips if the hash graphs have diverged.
Maybe we can continue using vector clocks on a best-effort basis, and
then use hashes to check consistency. This needs further thinking.