Strictly enforce canonical form of RLE-encoded buffers
This is useful when reconstructing changes from a compressed document.
Recomputing the hashes requires that the changes we construct are
bitwise identical to the original changes. Reconstructing the original
change is easier if there is only one way that a particular sequence of
values can be represented.

This means we don't want Postel's Law! If we are liberal in what inputs
we accept, we just create a headache when we need to reconstruct those
inputs later. By being strict about inputs, we remove ambiguity about
what the original input might have been.