Use (keyCtr, keyActor) = (0, null) for _head
Rather than (0, 0). The reason for this change is that when a change
gets applied to a document, an actor index of 0 in a change gets
transformed into the document's actor index for the author of that
change, which may be any number. This is confusing, since the head of a
list is not actually associated with any one particular actorId.