Use elemIds on the list editing actions
The list elemId (opId of the insertion operation that created the list
element) may be different from the opId of the current list element
value in the following scenario: a list element is created; then its
value is overwritten; then the document is saved and loaded again. The
patch that reconstitutes the loaded document will contain the correct
list elemIds in the insert actions.