Replace Frontend.getDeps with Backend.getHeads
The frontend does not always have a full picture of the latest change
hashes, since the binary encoding and hashing happens in the backend.
Therefore it makes more sense to have this function on the backend.

Putting the deps hashes in the backendState object rather than getting
them from the opSet because we sometimes (e.g. in Automerge.getChanges)
need to call getHeads on an old backend state object.