Fix deps for changes generated by frontend
When frontend and backend are on separate threads, and the frontend
generates several changes in quick succession without waiting for
response from the backend, then the frontend cannot fill in the deps as
required. Solved this by filling in the missing dep on the backend side.

The previous code did not correctly account for deps arising from remote
changes.