Move index of columns ahead of the column data
This shuffles around the order in which data appears in encoded changes
and documents. For a change, it makes very little difference, but for
large documents it may be useful: it ensures that at the beginning of a
file we first have a list of the columns and all their lengths, which is
then followed by the actual column data. Laying out the data this way
has the advantage that if we only want to read some of the columns, we
can calculate the file offset of the columns of interest directly from
the information at the beginning of the file, and then jump directly to
the offset in the file at which a particular column starts. It saves us
from having to read columns that we don't need.

Right now this doesn't matter much, since we operate on byte arrays that
are already fully loaded into memory, but I envisage sometime having a
file interface for loading documents that allows seeking to an offset.
It will be nice to be ready for that day when it comes.