Allow extra bytes at the end of a change
After the end of the last column, a change can now include arbitrary
additional data. This data is uninterpreted, but it it is included in
the length field, the checksum, and the dependency hash. Future versions
can use this facility to extend the data format without breaking
changes.

In the compressed document encoding, these extra bytes for each change
are saved so that we can reconstruct every original change, byte for
byte (necessary for making the dependency hashes work). In the common
case where there are no extra bytes, there is no additional overhead.