Distinguish between key and elemId in operations
Now the JSON representation of an operation in a change uses the `key`
property only when the key is a string (operations on a map or table
object), and the `elemId` property when the key is an operation ID
(operations on a list or text object). The special value '_head' (for
insertions at the head of a list) also uses `elemId`.

Previously, the process of converting a JSON change to binary relied on
the heuristic of whether the key contained an '@' symbol to determine
whether it was an elemId. This had the unfortunate side-effect of
preventing '@' symbols from occurring in the key of a map. This format
change allows us to remove that heuristic, so keys containing '@' are
now possible.