Rename chld columns to ref; remove link operation
Link operation is currently unused (it was previously there for undo,
but we have removed the undo feature for now). The chldActor and chldCtr
columns were only used by link operations, so they can be removed too.
However, for cursors we also need an (actor, counter) pair to reference
the list element in question, so I'm renaming these columns to refActor
and refCtr and will use them for cursor storage.