Add fast path to OpSet.getMissingChanges
Improves performance in the common case where user wants to just get the
most recent couple of changes, e.g. in the context of a frequent sync.
In this case, the cost becomes linear in the number of changes returned,
whereas previously it was linear in the size of the entire change
history.

In [this benchmark](https://gist.github.com/pvh/73473686865b9e421b620bd308fc9c58),
the case of 1,000 changes and syncing after each change speeds up by
roughly 10x after applying this patch.