Add opId field to SingleInsertEdit
I overlooked this in the original proposal. The reason we need this is
that if a new value is assigned to an existing list element, the patch
returned by `Backend.getPatch()` needs to include different IDs for the
elemId and the current value (the latter is needed for filling in pred).