Fix list edits when update occurs inside a conflicted list element
Changed it so that patches for both list and map objects initially have
a `props` property, which for lists is indexed by elemId. This makes it
easier and more robust to find the appropriate subpatch without having
to scan the list of edits. The `props` property is then deleted by
finalizePatch() before the patch is sent to the frontend.