Simplify block metadata
Make block.lastKey and block.numVisible single-valued properties
(previously they were maps from objectId to value). They now pertain
only to the object identified by block.lastObjectActor and
block.lastObjectCtr. This is possible because we never actually relied
on those properties for anything other than the last object in a block.

Moreover, this change fixes a bug that would cause root object
operations sometimes to be inserted in the wrong place because we
incorrectly skipped over the block that it was supposed to go into, on
the basis of looking at lastKey.