Fix out-of-order change application
There was a bug that meant that if applyChanges was called with changes
in an order that was inconsistent with the dependencies between the
changes, only some of the changes would get applied. Other changes would
remain in the queue even though they were ready to be applied. This PR
fixes the bug.